Ability to show drv files
Previously, passing a `.drv` file wouldn't work because that requires
passing `--derivation` flag to `nix path-info`. With this commit we
simply check if the given path ends with `.drv` and pass `--derivation`
accordingly.

This doesn't work if the given path is a symlink to a store path, but
we can handle that later.